requirements.txt
google-generativeai
python-dotenv
langchain
streamlit
PyPDF2
chromadb
faiss-cpu
langchain_google_genai